Cleanup tests and fix bug in frost-select
There was an `.eslintrc` in `tests/` which was defining some globals
which really shouldn't have been globals, so that was removed and the
appropriate `import` statements were added to affected tests.

There was also a bug stopping `frost-select` from allowing users to
type queries if the `selectedValue` was passed in but null
(which happens to be how `ember-frost-bunsen` uses it).

`eslint` was also updated to `3.x`